Move statistics Model to rocketchat:models
